\name{expr.norm}
\alias{expr.norm}
\docType{data}
\title{
The goldenspike gene expression data
}
\description{
The goldenspike gene expression data of 10a.dat.
}
\usage{data(expr.norm)}
\details{
It is the goldenspike data "10a.dat".
}
\source{
http://www2.ccr.buffalo.edu/halfon/spike/spikedownloads.html
}
\examples{
data(expr.norm)
}
\keyword{datasets}
